+++ Source/WebCore/platform/graphics/gstreamer/GStreamerUtilities.h
@@ -79,7 +79,7 @@ void mapGstBuffer(GstBuffer*);
 void unmapGstBuffer(GstBuffer*);
 bool initializeGStreamer();
 unsigned getGstPlayFlag(const char* nick);
-GstClockTime toGstClockTime(float time);
+GstClockTime toGstClockTime(double time);
 bool gstRegistryHasElementForMediaType(GList* elementFactories, const char* capsString);
 
 #if GST_CHECK_VERSION(1, 5, 3)
